Updated <supportedOn />
V0.2r1
In this update I have commited myself to making it a little more user
friendly if anyone wanted to add to this pack;

The Idea of adding as many versions of Windows that support ADMX was to
enable easy additions of Group Policy Templates for earlier versions of
Windows than 10.

I will be expanding this to other things too, such as MS Office and
Java.

In the mean time, whip it down and enjoy!